+++ examples/monitor.py
@@ -1,3 +1,4 @@
+#!/usr/bin/env python3
 #
 # Copyright (c) 2017-present, Facebook, Inc.
 # All rights reserved.
@@ -7,15 +8,10 @@
 # of patent rights can be found in the PATENTS file in the same directory.
 #
 
-from __future__ import absolute_import
-from __future__ import division
-from __future__ import print_function
-from __future__ import unicode_literals
-
 from itertools import chain
 from pprint import pprint
 
-from pystemd.dbuslib import apply_signature, DBus
+from pystemd.dbuslib import DBus, apply_signature
 
 
 def monitor(*args):