+++ pystemd/run.py
@@ -1,3 +1,4 @@
+#!/usr/bin/env python3
 #
 # Copyright (c) 2017-present, Facebook, Inc.
 # All rights reserved.
@@ -6,10 +7,6 @@
 # LICENSE file in the root directory of this source tree. An additional grant
 # of patent rights can be found in the PATENTS file in the same directory.
 #
-from __future__ import absolute_import
-from __future__ import division
-from __future__ import print_function
-from __future__ import unicode_literals
 
 import fcntl
 import os
@@ -22,7 +19,6 @@ import tty
 import uuid
 
 import pystemd
-
 from pystemd.dbuslib import DBus, DBusMachine
 from pystemd.exceptions import PystemdRunError
 from pystemd.systemd1 import Manager as SDManager, Unit