+++ tests/test_utils.py
@@ -1,3 +1,4 @@
+#!/usr/bin/env python3
 #
 # Copyright (c) 2017-present, Facebook, Inc.
 # All rights reserved.
@@ -7,12 +8,6 @@
 # of patent rights can be found in the PATENTS file in the same directory.
 #
 
-from __future__ import absolute_import
-from __future__ import division
-from __future__ import print_function
-from __future__ import unicode_literals
-
-import six
 from unittest import TestCase
 
 from pystemd.utils import x2char_star
@@ -20,12 +15,12 @@ from pystemd.utils import x2char_star
 
 class TestContextToCharStar(TestCase):
     def test_pass_normal_vars(self):
-        for elem in (0, six.b("hi"), True, [], {}, (3, 4), {3, 4}):
+        for elem in (0, b"hi", True, [], {}, (3, 4), {3, 4}):
             self.assertEqual(elem, x2char_star(elem))
 
     def test_convert_to_char(self):
         for elem in ("", "hi all"):
-            self.assertEqual(six.b(elem), x2char_star(elem))
+            self.assertEqual(elem.encode(), x2char_star(elem))
 
     def test_convert_all(self):
         self.assertEqual(b"true", x2char_star(True, convert_all=True))